requirements.txt
systemrdl-compiler>=1.4
jinja2
